vortexiwebsite/requirements.txt
flask
pillow
sqlalchemy
boto3
requests
Flask-WTF
Flask-APScheduler
redis
Flask-Limiter
Flask-SQLAlchemy
Flask-Cors
eyed3
ffmpeg-python
argon2-cffi
pyotp
python-slugify
cryptography
psycopg2-binary
pyqrcode
pypng
timeago
python-redis-lock